requirements-mac.txt
# AI Document Chatbot with RAG - Complete Requirements
# Last updated: June 25, 2025

# ===== CORE FRAMEWORK =====
streamlit>=1.28.0

# ===== DOCUMENT PROCESSING =====
pypdf2>=3.0.1              # PDF text extraction
python-docx>=0.8.11        # DOCX file processing
nltk>=3.8.1                # Natural language processing

# ===== VECTOR OPERATIONS & EMBEDDINGS =====
sentence-transformers>=2.2.2   # Semantic embeddings (all-MiniLM-L6-v2)
faiss-cpu>=1.7.4               # Vector similarity search
numpy>=1.24.0                  # Numerical operations
scikit-learn>=1.3.0            # TF-IDF fallback and ML utilities

# ===== API INTEGRATION =====
openai>=1.3.0              # OpenAI GPT-4o integration
groq>=0.4.1                # Groq LLaMA3-8B integration (primary LLM)

# ===== UTILITIES =====
python-dotenv>=1.0.0       # Environment variable management
pathlib>=1.0.1             # File path operations

# ===== OPTIONAL ENHANCEMENTS =====
# Uncomment these for enhanced functionality:

# Enhanced PDF Processing
# pymupdf>=1.23.0          # Better PDF text extraction (alternative to PyPDF2)

# Development Tools
# jupyter>=1.0.0           # Notebook development
# matplotlib>=3.7.0        # Data visualization
# pandas>=2.0.0            # Data manipulation

# Local Transformers (if needed)
# transformers>=4.33.0     # HuggingFace transformers
# torch>=2.0.0             # PyTorch backend

# Testing Framework
# pytest>=7.4.0            # Unit testing
# pytest-cov>=4.1.0        # Coverage testing